move helper-types to own file
